+++ testing/cffi1/buildtool_example/src/csrc/square.c
@@ -1,0 +1,5 @@
+#include "square.h"
+
+int square(int n) {
+    return n * n;
+}